requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
drf-yasg==1.21.7
python-decouple==3.8
Pillow==10.2.0
